move journal class files into journals dir
