+++ jrnl/journals/__init__.py
@@ -1,0 +1,5 @@
+from .DayOneJournal import DayOne
+from .Entry import Entry
+from .FolderJournal import Folder
+from .Journal import Journal
+from .Journal import open_journal
