+++ jrnl/plugins/dates_exporter.py
@@ -7,8 +7,8 @@ from typing import TYPE_CHECKING
 from jrnl.plugins.text_exporter import TextExporter
 
 if TYPE_CHECKING:
-    from jrnl.Entry import Entry
-    from jrnl.Journal import Journal
+    from jrnl.journals import Entry
+    from jrnl.journals import Journal
 
 
 class DatesExporter(TextExporter):
