+++ jrnl/plugins/fancy_exporter.py
@@ -13,8 +13,8 @@ from jrnl.messages import MsgText
 from jrnl.plugins.text_exporter import TextExporter
 
 if TYPE_CHECKING:
-    from jrnl.Entry import Entry
-    from jrnl.Journal import Journal
+    from jrnl.journals import Entry
+    from jrnl.journals import Journal
 
 
 class FancyExporter(TextExporter):
