+++ jrnl/plugins/jrnl_importer.py
@@ -11,7 +11,7 @@ from jrnl.messages import MsgText
 from jrnl.output import print_msg
 
 if TYPE_CHECKING:
-    from jrnl.Journal import Journal
+    from jrnl.journals import Journal
 
 
 class JRNLImporter:
