+++ jrnl/plugins/yaml_exporter.py
@@ -13,8 +13,8 @@ from jrnl.output import print_msg
 from jrnl.plugins.text_exporter import TextExporter
 
 if TYPE_CHECKING:
-    from jrnl.Entry import Entry
-    from jrnl.Journal import Journal
+    from jrnl.journals import Entry
+    from jrnl.journals import Journal
 
 
 class YAMLExporter(TextExporter):
@@ -34,7 +34,7 @@ class YAMLExporter(TextExporter):
         body = body_wrapper + entry.body
 
         tagsymbols = entry.journal.config["tagsymbols"]
-        # see also Entry.Entry.rag_regex
+        # see also Entry.rag_regex
         multi_tag_regex = re.compile(rf"(?u)^\s*([{tagsymbols}][-+*#/\w]+\s*)+$")
 
         """Increase heading levels in body text"""
